fix bug

+++ src/main/java/com/tele/goldenkey/controller/param/LiveUserParam.java
@@ -14,7 +14,7 @@ public class LiveUserParam {
 
     private Integer userId;
 
-    private Integer livedId;
+    private Long livedId;
 
     public LiveUserParam() {
     }
