+++ src/main/java/com/tele/goldenkey/live/AgreeMaiMaiEvent.java
@@ -22,7 +22,7 @@ public class AgreeMaiMaiEvent extends LiveEventCls {
     }
 
     @Override
-    public LiveEventDto getLiveEventDto(Integer livedId, Integer userId, Integer terminalId) throws ServiceException {
+    public LiveEventDto getLiveEventDto(Long livedId, Integer userId, Integer terminalId) throws ServiceException {
         ValidateUtils.notNull(terminalId);
         return new LiveEventDto(livedId, userId, terminalId);
     }
