+++ src/main/java/com/tele/goldenkey/live/CloseEvent.java
@@ -21,7 +21,7 @@ public class CloseEvent extends LiveEventCls {
     }
 
     @Override
-    public LiveEventDto getLiveEventDto(Integer livedId, Integer userId, Integer terminalId) throws ServiceException {
+    public LiveEventDto getLiveEventDto(Long livedId, Integer userId, Integer terminalId) throws ServiceException {
         ValidateUtils.notNull(livedId);
         return new LiveEventDto(livedId, null, null);
     }
