+++ src/main/java/com/tele/goldenkey/live/RefuseMaiEvent.java
@@ -18,7 +18,7 @@ public class RefuseMaiEvent extends LiveEventCls {
     }
 
     @Override
-    public LiveEventDto getLiveEventDto(Integer livedId, Integer userId, Integer terminalId) throws ServiceException {
+    public LiveEventDto getLiveEventDto(Long livedId, Integer userId, Integer terminalId) throws ServiceException {
         ValidateUtils.notNull(terminalId);
         return new LiveEventDto(livedId, userId, terminalId);
     }
